+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.x branch
 ## 2.17 branch
+### 2.17.6
+* widgets (cells) use collections;
+* added tests for `BackupsController` admin class.
+
 ### 2.17.5
 * added `GetStartAndEndDateTrait`;
 * added `LogsMethodsTrait`.